+++ src/Exporter.php
@@ -6,6 +6,8 @@ namespace Bakame\Aide\Profiler;
 
 interface Exporter
 {
+    public function exportSnapshot(Snapshot $snapshot): void;
+
     public function exportProfilingData(ProfilingResult|ProfilingData $profilingData): void;
 
     public function exportProfiler(Profiler $profiler): void;